+++ component/backend/Model/Releases.php
@@ -9,6 +9,7 @@ namespace Akeeba\ReleaseSystem\Admin\Model;
 
 defined('_JEXEC') or die;
 
+use Akeeba\ReleaseSystem\Admin\Model\Mixin\ClearCacheAfterActions;
 use FOF30\Container\Container;
 use FOF30\Model\DataModel;
 
@@ -78,6 +79,10 @@ class Releases extends DataModel
 	use Mixin\Assertions;
 	use Mixin\VersionedCopy {
 		Mixin\VersionedCopy::onBeforeCopy as onBeforeCopyVersioned;
+	}
+	use Mixin\ClearCacheAfterActions
+	{
+		Mixin\ClearCacheAfterActions::onAfterCopy as onAfterCopyCacheClean;
 	}
 
 	/** @var  DataModel\Collection  Used to handle copies */
@@ -280,7 +285,7 @@ class Releases extends DataModel
 		switch ($fltMaturity)
 		{
 			case 'beta':
-				$query->where($db->qn('maturity') . ' IN (' . $db->q('beta'), ',' . $db->q('rc') . ',' . $db->q('stable') . ')');
+				$query->where($db->qn('maturity') . ' IN (' . $db->q('beta') . ',' . $db->q('rc') . ',' . $db->q('stable') . ')');
 				break;
 			case 'rc':
 				$query->where($db->qn('maturity') . ' IN (' . $db->q('rc') . ',' . $db->q('stable') . ')');
@@ -665,5 +670,7 @@ class Releases extends DataModel
 		});
 
 		self::$itemsBeforeCopy = null;
+
+		$this->onAfterCopyCacheClean($releaseAfterCopy);
 	}
 }