+++ component/backend/View/wrongphp.php
@@ -5,6 +5,9 @@
  * @license   GNU General Public License version 3, or later
  */
 
+// Protect from unauthorized access
+defined('_JEXEC') or die();
+
 // PHP 5.6 is modern enough. Anything else gets a warning.
 $minPHPVersion = '5.4.0';
 $recommendedPHPVersion = '5.6.0';